requirements.txt
# Backend Dependencies for UNRC API

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# NLP & ML
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# File Handling
python-docx==0.8.11
pypdf==3.17.1

# CORS
python-cors==1.2.0

# Environment
python-dotenv==1.0.0

# Utilities
requests==2.31.0
typing-extensions==4.8.0
aiofiles==23.2.1

# Development (optional)
pytest==7.4.3
pytest-cov==4.1.0

# Nota: El modelo spaCy en español se descarga automáticamente
# Ejecuta después: python -m spacy download es_core_news_sm